requirements.txt
appdirs==1.4.3
autopep8==1.4.4
bootstrap4==0.1.0
CacheControl==0.12.6
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2.14
django-allauth==0.39.1
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.3.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-tinymce==3.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
html5lib==1.0.1
idna==2.8
instamojo-wrapper==1.2.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.0.1
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==6.2.2
progress==1.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-decouple==3.1
python3-openid==3.1.0
pytoml==0.1.21
pytz==2018.5
razorpay==1.2.0
reportlab==3.5.51
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
six==1.14.0
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2
webencodings==0.5.1
xhtml2pdf==0.2.4
